+++ samples/music/src/main/java/org/jboss/elemento/sample/music/Track.java
@@ -1,0 +1,39 @@
+/*
+ *  Copyright 2023 Red Hat
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package org.jboss.elemento.sample.music;
+
+import java.util.List;
+
+import elemental2.core.JsArray;
+import jsinterop.annotations.JsOverlay;
+import jsinterop.annotations.JsPackage;
+import jsinterop.annotations.JsType;
+
+import static java.util.Collections.emptyList;
+
+@JsType(isNative = true, namespace = JsPackage.GLOBAL, name = "Object")
+public class Track {
+
+    public int track;
+    public String title;
+    public JsArray<String> writer;
+    public String length;
+
+    @JsOverlay
+    public final List<String> writer() {
+        return writer == null ? emptyList() : writer.asList();
+    }
+}